+++ src/gocept/selenium/webdriver.py
@@ -18,9 +18,8 @@ import gocept.selenium.wd_selenese
 import os
 import os.path
 import selenium.webdriver
-import shutil
 import sys
-import tempfile
+import warnings
 
 
 # work around Python 2.4 lack of absolute_import,
@@ -30,27 +29,44 @@ plonetesting = __import__('plone.testing', {}, {}, [''])
 
 class Layer(plonetesting.Layer):
 
-    _browser = os.environ.get('GOCEPT_WEBDRIVER_BROWSER', 'firefox')
+    profile = None
+    headless = False
+    _browser = 'firefox'
 
     def setUp(self):
         if 'http_address' not in self:
             raise KeyError("No base layer has set self['http_address']")
 
-        if self._browser.lower() == 'firefox':
+        browser = os.environ.get('GOCEPT_WEBDRIVER_BROWSER')
+        headless = os.environ.get('GOCEPT_SELENIUM_HEADLESS')
+
+        if headless is None or headless.lower() not in ['true', 'false']:
+            warnings.warn('GOCEPT_SELENIUM_HEADLESS invalid. \
+                          Possible Values are true and false. Got: %s.\
+                          Falling back to default (false).' %
+                          os.environ.get('GOCEPT_SELENIUM_HEADLESS'))
+
+        if headless.lower() == 'true':
+            self.headless = True
+
+        if browser is None or browser.lower() not in ['chrome', 'firefox']:
+            warnings.warn('GOCEPT_WEBDRIVER_BROWSER invalid. \
+                          Possible Values are firefox and chrome. Got: %s.\
+                          Falling back to firefox.' %
+                          os.environ.get('GOCEPT_WEBDRIVER_BROWSER'))
+
+        if browser.lower() == 'chrome':
+            self._browser = 'chrome'
+        else:
             self.profile = FirefoxProfile(
                 os.environ.get(
                     'GOCEPT_WEBDRIVER_FF_PROFILE',
                     os.environ.get('GOCEPT_SELENIUM_FF_PROFILE')))
             self.profile.native_events_enabled = True
             self.profile.update_preferences()
-        else:
-            self.profile = None
 
         self._start_selenium()
         atexit.register(self._stop_selenium)
-        speed = os.environ.get('GOCEPT_SELENIUM_SPEED')
-        if speed is not None:
-            self['seleniumrc'].setSpeed(speed)
 
     def tearDown(self):
         self._stop_selenium()
@@ -59,9 +75,28 @@ class Layer(plonetesting.Layer):
         del self['seleniumrc']
 
     def _start_selenium(self):
-        options = selenium.webdriver.FirefoxOptions()
-        self['seleniumrc'] = selenium.webdriver.Firefox(
-            firefox_profile=self.profile, options=options)
+        if self._browser == 'firefox':
+            options = selenium.webdriver.FirefoxOptions()
+
+            if self.headless:
+                options.add_argument('-headless')
+
+            self['seleniumrc'] = selenium.webdriver.Firefox(
+                firefox_profile=self.profile, options=options)
+
+        if self._browser == 'chrome':
+            options = selenium.webdriver.ChromeOptions()
+            options.add_argument('--disable-dev-shm-usage')
+
+            if self.headless:
+                options.add_argument('--headless')
+            else:
+                raise NotImplementedError(
+                    'Chromedriver currently only works headless.')
+
+            self['seleniumrc'] = selenium.webdriver.Chrome(
+                options=options,
+                service_args=['--log-path=chromedriver.log'])
 
     def _stop_selenium(self):
         # Only stop selenium if it is still active.
@@ -69,24 +104,8 @@ class Layer(plonetesting.Layer):
                 self['seleniumrc'].session_id is None):
             return
 
+        # Quit also removes the tempdir the ff profile is copied in.
         self['seleniumrc'].quit()
-
-        if self.profile and os.path.exists(self.profile.profile_dir):
-            shutil.rmtree(self.profile.profile_dir)
-            if os.path.dirname(
-                    self.profile.profile_dir) != tempfile.gettempdir():
-                try:
-                    os.rmdir(os.path.dirname(self.profile.profile_dir))
-                except OSError:
-                    pass
-
-            if not os.environ.get('GOCEPT_SELENIUM_KEEP_NATIVE_FF_EVENTS_LOG'):
-                native_ff_events_log = os.path.join(
-                    tempfile.gettempdir(), 'native_ff_events_log')
-                try:
-                    os.remove(native_ff_events_log)
-                except OSError:
-                    pass
 
 
 class WebdriverSeleneseLayer(plonetesting.Layer):